Bump RC, firesim and barstools for chisel3.4 updates
Note: firesim and barstools point to commits in the sifive forks of those repos
I didn't update the URL in .gitmodules because I'm not sure how that works in a PR
(because you wouldn't really want to merge sync'ing to the sifive repo).

Requires: ucb-bar/barstools#92 and firesim/firesim#658

The version of rocket-chip, chisel3 and firrtl is chosen here because it is
the latest known to pass my tests.  You will likely want to bump further.